% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_engine.R
\name{.generate_population_cpp}
\alias{.generate_population_cpp}
\title{Generate a population using the C++ engine (internal)}
\usage{
.generate_population_cpp(
  n,
  seed,
  state,
  city,
  gender,
  min_age,
  max_age,
  modules,
  end_date,
  mc.cores,
  cpp_modules
)
}
\description{
Generate a population using the C++ engine (internal)
}
\keyword{internal}
