% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.get_cpp_modules}
\alias{.get_cpp_modules}
\title{Return cached compiled C++ modules, compiling them on first call}
\usage{
.get_cpp_modules(modules)
}
\description{
Return cached compiled C++ modules, compiling them on first call
}
\keyword{internal}
